chore(claude-sonnet-4-5): if a git pull request is made on one of the agent's work,...
